traditional and json service added

+++ EFCoreWithPostgreSQL/Models/OrderDetails/OrderDetailEntity.cs
@@ -7,9 +7,9 @@ namespace EFCoreJsonApp.Models.OrderDetails
     {
         public Guid OrderId { get; set; }
         public string ItemName { get; set; }
-        public double Price { get; set; }
+        public float Price { get; set; }
         public int Quantity { get; set; }
-        public double Total { get; set; }
+        public float Total { get; set; }
         public OrderEntity Order { get; set; }
     }
 }
